requirements.txt
google-generativeai
pandas
requests
beautifulsoup4
sqlalchemy